[PI][OpenCL] Test support for exposing compute slices
This test tries to identify devices at three levels - root, sub, and sub-sub. For a device at a given level, checks are provided to see if there is more than 1 device at its sub-level. This test is expected to run without crashing for all backends, though it is specifically added to test changes in OpenCL plugin.

Signed-off-by: Arvind Sudarsanam <arvind.sudarsanam@intel.com>